Use avy-goto-char instead of avy-goto-char-2
Most of the time, it is seldom for the selection of a single character
to take more than 2 specifiers afterwards. With avy-goto-char-2, a
specifier is typically still required. So the same amount of key
presses most of the time. But avy-goto-char makes it easier to point
to where the cursor should jump to. E.g. the last character of a line.